refactor(agent): extract the queued-run failure recovery from AgentRuntime
What becomes of a queued run that failed moves into
QueuedRunFailureRecovery.

Four conditions can each independently force a dead-letter rather than a
retry: an error class retrying cannot fix, a non-idempotent write fenced
in flight, an exhausted requeue budget, and a lost claim on the row. The
order between them is the behaviour, and it reads better in a class named
after the decision than inside the run loop it interrupts.

The retry budget and the backoff stay AgentRuntime constants — the reaper
command and the playground read them as published policy — so only the
decision that consumes them moved. The collaborator is optional in the
constructor for the same reason the ones above it are, and a null builds
one from the persister, logger and bus the runtime already holds.
